Add support for ignoring specified files or directories

+++ turbovisor.lua
@@ -21,8 +21,9 @@ local turbo = require "turbo"
 --- Parsing arguments for turbovisor
 -- @param arg All command line input. Note arg[0] is 'turbovisor', arg[1] is
 --    application name; so user-defined argument starts from arg[2]
--- @return a table containing option/value pair, e.g.
---    1. {"watch" = {".", "turbo/examples"}, ......}
+-- @return a table containing option/value pair, options including:
+--    'watch': set of files or directories to watch
+--    'ignore': set of files or directories that turbovisor shouldn't watch
 local function get_param(arg)
     local arg_opt
     local arg_tbl = {}
@@ -30,10 +31,13 @@ local function get_param(arg)
         if arg[i] == '--watch' or arg[i] == '-w' then
             arg_tbl.watch = {}
             arg_opt = 'watch'
-        elseif arg[i] == '--ignore' or arg[i] == '-i' then -- NYI
+        elseif arg[i] == '--ignore' or arg[i] == '-i' then
             arg_tbl.ignore = {}
             arg_opt = 'ignore'
         else
+            if string.sub(arg[i], 1, 2) == "./" then
+                arg[i] = string.sub(arg[i], 3) -- pass './'
+            end
             table.insert(arg_tbl[arg_opt], arg[i])
         end
     end
@@ -50,17 +54,18 @@ local turbovisor = class("turbovisor", turbo.ioloop.IOLoop)
 --- Start supervising.
 function turbovisor:supervise()
     -- Get command line parameters
-    local arg_tbl = get_param(arg)
+    self.arg_tbl = get_param(arg)
     -- Create a new inotify
-    local i_fd = turbo.inotify:new()
+    self.i_fd = turbo.inotify:new()
     -- Create a buffer for reading event in callback handler
     self.buf = ffi.gc(ffi.C.malloc(turbo.fs.PATH_MAX), ffi.C.free)
-    -- Initialize ioloop, add inotify handler and watched target
+    -- Initialize ioloop, add inotify handler
     self:initialize()
-    self:add_handler(i_fd, turbo.ioloop.READ, self.restart, self)
-    for i, target in pairs(arg_tbl.watch) do
+    self:add_handler(self.i_fd, turbo.ioloop.READ, self.restart, self)
+    -- Set watch on target file or directory
+    for i, target in pairs(self.arg_tbl.watch) do
         if turbo.fs.is_dir(target) then
-            turbo.inotify:watch_all(target)
+            turbo.inotify:watch_all(target, self.arg_tbl.ignore)
         else
             turbo.inotify:watch_file(target)
         end
@@ -89,15 +94,20 @@ function turbovisor.restart(self, fd, events)
     -- Read out event
     ffi.C.read(fd, self.buf, turbo.fs.PATH_MAX);
     self.buf = ffi.cast("struct inotify_event*", self.buf)
-    if self.buf.len ~= 0 then
-        turbo.log.notice("[turbovisor.lua] File "
-                             .. ffi.string(self.buf.name) ..
-                             " changed, application restarted!")
-    else                     -- 'len = 0' if we watch on file directly
-        turbo.log.notice("[turbovisor.lua] File "
-                             .. turbo.inotify:get_watched_file(self.buf.wd) ..
-                             " changed, application restarted!")
+    local full_path
+    if self.buf.len == 0 then -- 'len = 0' if we watch on file directly
+        full_path = turbo.inotify:get_watched_file(self.buf.wd)
+    else
+        local path = turbo.inotify:get_watched_file(self.buf.wd)
+        full_path = path .. '/' .. ffi.string(self.buf.name)
+        if path == '.' then full_path = ffi.string(self.buf.name) end
     end
+    -- Simply return if we need to ignore the file
+    if turbo.util.is_in(full_path, self.arg_tbl.ignore) then
+        return
+    end
+    turbo.log.notice("[turbovisor.lua] File '" .. full_path ..
+                         "' changed, application restarted!")
     -- Restart application
     ffi.C.kill(self.cpid, 9)
     ffi.C.waitpid(self.cpid, status, 0)
